completed through 3.7.3.1. One difference between axum and actix is that a parsing error (e.g. Deserialization of a form fails) returns a 422 error code in axum
This seems to be fine, and more correct than a 400 return code; it is
emitted because the form could be parsed but there are missing fields.

Something to check with the axum folks is if an html form is malformed
(meta questiwhat does that look like?) is the correct error emitted? my guess would
be a 400 would be most apropriate